stack using a linked list
This program implements a stack using a linked list in C++, featuring basic stack operations:

Push: Adds a new element to the top of the stack.
Pop: Removes the top element from the stack and displays it. If the stack is empty, a message is shown.
Display: Lists all elements currently in the stack from top to bottom.
getTop: Shows the current top element without removing it.